+++ src/test/java/com/threecats/appirater/prefs/FakePrefs.kt
@@ -1,8 +1,8 @@
 package com.threecats.appirater.prefs
 
 class FakePrefs(
-        override var eventCount: Long = 0,
-        override var launchCount: Long = 0,
+        override var eventCount: Int = 0,
+        override var launchCount: Int = 0,
         override var dateFirstLaunch: Long = 0,
         override var dontRate: Boolean = false,
         override var doneRate: Boolean = false,
